+++ src/SQLProvider/SqlRuntime.QueryExpression.fs
@@ -28,10 +28,17 @@ module internal QueryExpressionTransformer =
                 Some ((al,coltyp,rtyp), exp, oper.Type)
             | _ -> None
 
+        let projectionMap = Dictionary<string,ProjectionParameter ResizeArray>()
+        let groupProjectionMap = ResizeArray<alias*SqlColumnType>()
+
         let (|SingleTable|MultipleTables|) = function
             | MethodCall(None, MethodWithName "Select", [Constant(_, t) ;exp]) when t = typeof<System.Linq.IQueryable<SqlEntity>> || t = typeof<System.Linq.IOrderedQueryable<SqlEntity>> ->
                 SingleTable exp
             | MethodCall(None, MethodWithName "Select", [_ ;exp]) ->
+                MultipleTables exp
+            | Lambda([ParamName sourceAlias],(SqlColumnGet(entity,ct,rtyp) as oper)) as exp ->
+                if not(groupProjectionMap.Contains (entity,ct)) then
+                    groupProjectionMap.Add(entity,ct)
                 MultipleTables exp
             | _ -> failwith ("Unsupported projection: " + projection.NodeType.ToString())
 
@@ -42,9 +49,6 @@ module internal QueryExpressionTransformer =
             //   previously selected individual columns.
             // in the second case we also need to change any property on the input tuple into calls
             // onto GetSubEntity on the result parameter with the correct alias
-
-        let projectionMap = Dictionary<string,ProjectionParameter ResizeArray>()
-        let groupProjectionMap = ResizeArray<SqlColumnType>()
         
         let (|SourceTupleGet|_|) (e:Expression) =
             match e with
@@ -71,7 +75,7 @@ module internal QueryExpressionTransformer =
                     if aliasEntityDict.ContainsKey(alias) then
                         Some (alias,aliasEntityDict.[alias].FullName, None)
                     elif ultimateChild.IsSome then
-                        Some (alias,snd(ultimateChild.Value).FullName, None)
+                        Some (fst(ultimateChild.Value),snd(ultimateChild.Value).FullName, None)
                     else None
                 else None
             | PropertyGet(Some(PropertyGet(Some(ParamWithName "tupledArg"), nestedTuple)), info) 
@@ -108,7 +112,10 @@ module internal QueryExpressionTransformer =
                     decimalTypes  |> Seq.exists(fun t -> t = ty) || integerTypes |> Seq.exists(fun t -> t = ty)
 
                 let op =
-                    if me.Arguments.Count = 1 && (me.Arguments.[0].NodeType = ExpressionType.Parameter || (me.Arguments.[0].NodeType = ExpressionType.New && me.Arguments.[0].Type.Name.StartsWith("Grouping"))) then
+                    if me.Arguments.Count = 1 && (me.Arguments.[0].NodeType = ExpressionType.Parameter ||
+                                                    (me.Arguments.[0].NodeType = ExpressionType.New && me.Arguments.[0].Type.Name.StartsWith("Grouping")) || 
+                                                    (me.Arguments.[0].NodeType = ExpressionType.MemberAccess && me.Arguments.[0].Type.Name.StartsWith("IGrouping"))
+                                                 ) then
                         match me.Method.Name with
                         | "Count" -> Some (CountOp "", None)
                         | "Sum" when isNumType me.Type -> Some (SumOp "", None)
@@ -122,20 +129,23 @@ module internal QueryExpressionTransformer =
                     elif me.Arguments.Count = 2 then
                         match me.Arguments.[1] with
                         | :? LambdaExpression as la ->
+
+                            let getOp al op =
+                                let key = Utilities.getBaseColumnName op
+                                match me.Method.Name with
+                                | "Count" -> Some (CountOp key, Some (al,op))
+                                | "Sum" -> Some (SumOp key, Some (al,op))
+                                | "Avg" | "Average" -> Some (AvgOp key, Some (al,op))
+                                | "Min" -> Some (MinOp key, Some (al,op))
+                                | "Max" -> Some (MaxOp key, Some (al,op))
+                                | "StdDev" | "StDev" | "StandardDeviation" -> Some (StdDevOp key, Some (al,op))
+                                | "Variance" -> Some (VarianceOp key, Some (al,op))
+                                | _ -> None
+
                             let rec directAggregate (exp:Expression) =
                                 match exp.NodeType, exp with
                                 | _, OptionalConvertOrTypeAs(SqlColumnGet(entity, op, _)) ->
-                                    let key = Utilities.getBaseColumnName op
-
-                                    match me.Method.Name with
-                                    | "Count" -> Some (CountOp key, Some op)
-                                    | "Sum" -> Some (SumOp key, Some op)
-                                    | "Avg" | "Average" -> Some (AvgOp key, Some op)
-                                    | "Min" -> Some (MinOp key, Some op)
-                                    | "Max" -> Some (MaxOp key, Some op)
-                                    | "StdDev" | "StDev" | "StandardDeviation" -> Some (StdDevOp key, Some op)
-                                    | "Variance" -> Some (VarianceOp key, Some op)
-                                    | _ -> None
+                                    getOp entity op
                                 | ExpressionType.Quote, (:? UnaryExpression as ce) 
                                 | ExpressionType.Convert, (:? UnaryExpression as ce) -> directAggregate ce.Operand
                                 | ExpressionType.MemberAccess, ( :? MemberExpression as me2) -> 
@@ -144,6 +154,14 @@ module internal QueryExpressionTransformer =
                                     | _ -> None
                                 // This lamda could be parsed more if we would want to support
                                 // more complex aggregate scenarios.
+                                // Subtables
+                                | _, OptionalFSharpOptionValue(MethodCall(Some(o),((MethodWithName "GetColumn" as meth) | (MethodWithName "GetColumnOption" as meth)),[String key])) when o.Type.Name = "SqlEntity" -> 
+                                    match o.NodeType, o with
+                                    | ExpressionType.Call, (:? MethodCallExpression as ce)
+                                            when (ce.Method.Name = "GetSubTable" && ce.Object <> null && ce.Object :? ParameterExpression ) ->
+                                        let alias = (ce.Object :?> ParameterExpression).Name
+                                        getOp alias (KeyColumn key)
+                                    | _ -> None
                                 | _ -> None
                             directAggregate la.Body
                         | _ -> None
@@ -158,16 +176,21 @@ module internal QueryExpressionTransformer =
                             | KeyOp x -> x
                     //Count 1 is over all the items
                     let vf = if v = "" then None else Some v
-                    let ty = typedefof<GroupResultItems<_>>.MakeGenericType(me.Arguments.[0].Type.GetGenericArguments().[0])
+                    let ty =
+                        let retType =
+                            if me.Arguments.[0].Type.GetGenericArguments().Length > 1 then
+                                me.Arguments.[0].Type.GetGenericArguments().[1]
+                            else typeof<SqlEntity>
+                        typedefof<GroupResultItems<_,_>>.MakeGenericType(me.Arguments.[0].Type.GetGenericArguments().[0], retType)
                     let aggregateColumn = Expression.Constant(vf, typeof<Option<string>>) :> Expression
                     let meth = ty.GetMethod(methodname)
                     let generic = meth.MakeGenericMethod(me.Method.ReturnType);
-                    let replacementExpr = 
+                    let replacementExpr =
                         Expression.Call(Expression.Convert(me.Arguments.[0], ty), generic, aggregateColumn)
                     let res =
                         match calcs with
-                        | None -> Some (GroupColumn(o,SqlColumnType.KeyColumn(v)), replacementExpr)
-                        | Some calculation -> Some (GroupColumn(o,calculation), replacementExpr)
+                        | None -> Some (("",GroupColumn(o,SqlColumnType.KeyColumn(v))), replacementExpr)
+                        | Some (al,calculation) -> Some ((al,GroupColumn(o,calculation)), replacementExpr)
                     if res.IsSome && groupProjectionMap.Contains(fst(res.Value)) then None
                     else res
                 | _ -> None
@@ -226,9 +249,15 @@ module internal QueryExpressionTransformer =
                 | false, _ -> projectionMap.Add(alias,new ResizeArray<_>(seq{yield EntityColumn(key)}))
                 | _ -> ()
                 Some
-                    (Expression.Call(
-                        Expression.Call(databaseParam,getSubEntityMi,Expression.Constant(alias),Expression.Constant(name)),
-                            mi,Expression.Constant(key)))
+                    (match databaseParam.Type.Name.StartsWith("IGrouping") with
+                     | false ->
+                        (Expression.Call(
+                            Expression.Call(databaseParam,getSubEntityMi,Expression.Constant(alias),Expression.Constant(name)),
+                                mi,Expression.Constant(key)))
+                     | true ->
+                        (Expression.Call(
+                            Expression.Call(Expression.Parameter(typeof<SqlEntity>,alias),getSubEntityMi,Expression.Constant(alias),Expression.Constant(name)),
+                                mi,Expression.Constant(key))))
 
             | ExpressionType.Call, MethodCall(Some(ParamName pname as p),(MethodWithName "GetColumn" | MethodWithName "GetColumnOption" as mi),[String key]) 
                     when p.Type = typeof<SqlEntity> ->
@@ -330,7 +359,8 @@ module internal QueryExpressionTransformer =
             | ExpressionType.Call,               (:? MethodCallExpression as e)  -> let transformed = Expression.Call( (if e.Object = null then null else transform en e.Object), e.Method, e.Arguments |> Seq.map(fun a -> transform en a))
                                                                                     match transformed with
                                                                                     | GroupByAggregate(param, callreplace) -> 
-                                                                                        groupProjectionMap.Add(param)
+                                                                                        if not(groupProjectionMap.Contains param) then
+                                                                                            groupProjectionMap.Add(param)
                                                                                         upcast callreplace
                                                                                     | _ -> 
                                                                                         upcast transformed
@@ -406,9 +436,12 @@ module internal QueryExpressionTransformer =
                 // select x
                 // this does not create a call to .select() after .where(), therefore in this case we must provide our own projection that simply selects a whole row
                 let initDbParam = 
-                    match sqlQuery.Grouping.Length > 0 with
-                    | true -> Expression.Parameter(typeof<System.Linq.IGrouping<_,SqlEntity>>,"result")
-                    | false -> Expression.Parameter(typeof<SqlEntity>,"result")
+                    match sqlQuery.Grouping.Length > 0, sqlQuery.Links.Length with
+                    | true, 0 -> Expression.Parameter(typeof<System.Linq.IGrouping<_,SqlEntity>>,"result")
+                    | true, 1 -> Expression.Parameter(typeof<System.Linq.IGrouping<_,Microsoft.FSharp.Linq.RuntimeHelpers.AnonymousObject<SqlEntity,SqlEntity>>>,"result")
+                    | true, 2 -> Expression.Parameter(typeof<System.Linq.IGrouping<_,Microsoft.FSharp.Linq.RuntimeHelpers.AnonymousObject<SqlEntity,SqlEntity,SqlEntity>>>,"result")
+                    | true, 3 -> Expression.Parameter(typeof<System.Linq.IGrouping<_,Microsoft.FSharp.Linq.RuntimeHelpers.AnonymousObject<SqlEntity,SqlEntity,SqlEntity,SqlEntity>>>,"result")
+                    | _ -> Expression.Parameter(typeof<SqlEntity>,"result")
                 let pmap = Dictionary<string,ProjectionParameter ResizeArray>()
                 pmap.Add(baseAlias, new ResizeArray<_>())
                 (Expression.Lambda(initDbParam,initDbParam).Compile(),pmap)
@@ -505,25 +538,26 @@ module internal QueryExpressionTransformer =
                                 // GroupBy: collect aggreagte operations
                                 // Should gather the column names what we want to aggregate, not just operations.
                                 let aggregations:(alias * SqlColumnType) list =
-                                    groupProjectionMap 
+                                    List.concat [ x; (groupProjectionMap |> Seq.toList)]
                                     |> Seq.map(fun op ->
                                         if not (group |> List.isEmpty) then 
                                             group 
-                                            |> List.choose(fun (a, cc) -> 
+                                            |> List.choose(fun (baseal, cc) -> 
                                                 match baseKey(cc), op with
-                                                | KeyColumn c, GroupColumn(AvgOp "", KeyColumn "") -> Some (a, GroupColumn(AvgOp c, KeyColumn c))
-                                                | KeyColumn c, GroupColumn(MinOp "", KeyColumn "") -> Some (a, GroupColumn(MinOp c, KeyColumn c))
-                                                | KeyColumn c, GroupColumn(MaxOp "", KeyColumn "") -> Some (a, GroupColumn(MaxOp c, KeyColumn c))
-                                                | KeyColumn c, GroupColumn(SumOp "", KeyColumn "") -> Some (a, GroupColumn(SumOp c, KeyColumn c))
-                                                | KeyColumn c, GroupColumn(StdDevOp "", KeyColumn "") -> Some (a, GroupColumn(StdDevOp c, KeyColumn c))
-                                                | KeyColumn c, GroupColumn(VarianceOp "", KeyColumn "") -> Some (a, GroupColumn(VarianceOp c, KeyColumn c))
-                                                | KeyColumn c, GroupColumn(KeyOp "", KeyColumn "") -> Some (a, GroupColumn(KeyOp c, KeyColumn c))
-                                                | KeyColumn c, GroupColumn(CountOp "", KeyColumn "") -> Some (a, GroupColumn(CountOp c, KeyColumn c))
-                                                | KeyColumn c, GroupColumn(agg, KeyColumn g) when g <> "" -> Some (a, op)
-                                                | KeyColumn c, GroupColumn(_) when Utilities.getBaseColumnName op <> "" -> Some (a, op)
+                                                | KeyColumn c, (a,GroupColumn(AvgOp "", KeyColumn "")) -> Some (a, GroupColumn(AvgOp c, KeyColumn c))
+                                                | KeyColumn c, (a,GroupColumn(MinOp "", KeyColumn "")) -> Some (a, GroupColumn(MinOp c, KeyColumn c))
+                                                | KeyColumn c, (a,GroupColumn(MaxOp "", KeyColumn "")) -> Some (a, GroupColumn(MaxOp c, KeyColumn c))
+                                                | KeyColumn c, (a,GroupColumn(SumOp "", KeyColumn "")) -> Some (a, GroupColumn(SumOp c, KeyColumn c))
+                                                | KeyColumn c, (a,GroupColumn(StdDevOp "", KeyColumn "")) -> Some (a, GroupColumn(StdDevOp c, KeyColumn c))
+                                                | KeyColumn c, (a,GroupColumn(VarianceOp "", KeyColumn "")) -> Some (a, GroupColumn(VarianceOp c, KeyColumn c))
+                                                | KeyColumn c, (a,GroupColumn(KeyOp "", KeyColumn "")) -> Some (a, GroupColumn(KeyOp c, KeyColumn c))
+                                                | KeyColumn c, (a,GroupColumn(CountOp "", KeyColumn "")) -> Some (a, GroupColumn(CountOp c, KeyColumn c))
+                                                | KeyColumn c, (a,GroupColumn(agg, KeyColumn g)) when g <> "" -> Some (op)
+                                                | KeyColumn c, (a,GroupColumn(_)) when Utilities.getBaseColumnName (snd op) <> "" -> Some (op)
+                                                | KeyColumn c, (a,KeyColumn(c2)) when Utilities.getBaseColumnName (snd op) <> "" -> Some (op)
                                                 | _ -> None)
 
-                                        else ["",op]
+                                        else [op]
                                     ) |> Seq.concat |> Seq.toList
                                 group, aggregations)
                         groupgin.AddRange(gatheredAggregations)
@@ -545,7 +579,12 @@ module internal QueryExpressionTransformer =
 
         let sqlQuery = 
             if groupgin.Count > 0 then
-                { sqlQuery with Grouping = groupgin |> Seq.toList }
+                let cleaned =
+                    let lst = groupgin |> Seq.toList
+                    match lst |> List.filter(fun (keys,agg) -> not agg.IsEmpty) with
+                    | [] -> lst
+                    | filtered -> filtered
+                { sqlQuery with Grouping = cleaned |> Seq.distinct |> Seq.toList }
             else sqlQuery
 
         // a special case here to handle queries that start from the relationship of an individual
@@ -558,6 +597,18 @@ module internal QueryExpressionTransformer =
                    | x -> // Multiple aliases for same basetable name. We have to select one and merge columns.
                         let starSelect = projectionColumns |> Seq.tryPick(fun p -> match p.Value.Count with | 0 -> Some p | _ -> None)
                         match starSelect with
+                        | Some sel when useCanonicalsOnSelect && projectionColumns |> Seq.exists(fun kvp ->
+                                        kvp.Value |> Seq.exists(fun co -> match co with | OperationColumn _ -> true | EntityColumn _ -> false)) ->
+                            // Whole entity plus operation columns. We need urgent table lookup.
+                            let myLock = provider.GetLockObject()
+                            let cols = lock myLock (fun () -> provider.GetColumns (con,snd sqlQuery.UltimateChild.Value))
+                            let opcols = projectionColumns |> Seq.map(fun p -> p.Value) |> Seq.concat
+                            let allcols = cols |> Map.toSeq |> Seq.map(fun (k,_) -> EntityColumn k)
+                            let itms = Seq.concat [|opcols;allcols|] |> Seq.distinct |> Seq.toList
+                            let tmp = sel
+                            projectionColumns.Clear()
+                            projectionColumns.Add(tmp.Key, new ResizeArray<ProjectionParameter>(itms))
+                            tmp.Key
                         | Some sel ->
                             let tmp = sel
                             projectionColumns.Clear()
@@ -573,6 +624,7 @@ module internal QueryExpressionTransformer =
                { sqlQuery with UltimateChild = Some(alias,snd sqlQuery.UltimateChild.Value) }, alias
             else sqlQuery,baseAlias
 
+        let baseName = if baseAlias <> "" then baseAlias else (fst sqlQuery.UltimateChild.Value)
         let resolve defaultTable name =
             // name will be blank when there is only a single table as it never gets
             // tupled by the LINQ infrastructure. In this case we know it must be referring
@@ -580,10 +632,10 @@ module internal QueryExpressionTransformer =
             if String.IsNullOrWhiteSpace(name) || name = "__base__" || entityIndex.Count = 0 then
                 match defaultTable with
                 | Some(s) -> s
-                | None -> (fst sqlQuery.UltimateChild.Value)
+                | None -> baseName
             else 
                 let tbl = Utilities.resolveTuplePropertyName name entityIndex
-                if tbl = "" then baseAlias else tbl
+                if tbl = "" then baseName else tbl
 
         let resolve name =
             // name will be blank when there is only a single table as it never gets
@@ -592,7 +644,7 @@ module internal QueryExpressionTransformer =
             if String.IsNullOrWhiteSpace(name) || name = "__base__" || entityIndex.Count = 0 then (fst sqlQuery.UltimateChild.Value)
             else 
                 let tbl = Utilities.resolveTuplePropertyName name entityIndex
-                if tbl = "" then baseAlias else tbl
+                if tbl = "" || not(entityIndex.Contains tbl) then baseName else tbl
 
         
         // Resolves aliases on canonical multi-column functions
@@ -667,7 +719,10 @@ module internal QueryExpressionTransformer =
         // much easier at this stage once we have knowledge of the whole query
         let sqlQuery = { sqlQuery with Filters = List.map resolveFilterList sqlQuery.Filters
                                        Ordering = List.map(function 
-                                                            |("",b,c) -> (resolve "",resolveC b,c) 
+                                                            |("",b,c) ->
+                                                                match resolve "" with
+                                                                | "" when baseAlias <> "" -> (baseAlias,resolveC b,c) 
+                                                                | x -> (x,resolveC b,c) 
                                                             | a,c,b -> resolve a, resolveC c,b) sqlQuery.Ordering 
                                        // Resolve other canonical function columns also:
                                        Grouping = sqlQuery.Grouping |> List.map(fun (g,a) -> g|>List.map(fun (ga, gk) -> resolve ga, resolveC gk), a|>List.map(fun(ag,aa)->resolve ag,resolveC aa)) 
@@ -715,7 +770,9 @@ module internal QueryExpressionTransformer =
                     }
 
             if linkData.ForeignTable.Name <> "" then (outerAlias, resolvedLinkData, innerAlias)
-            else (outerAlias, { resolvedLinkData with ForeignTable = resolved }, innerAlias)
+            else
+                let oa = if outerAlias = "" then resolved.Name else outerAlias
+                (oa, { resolvedLinkData with ForeignTable = resolved }, innerAlias)
 
         let sqlQuery = { sqlQuery with Links = List.map resolveLinks sqlQuery.Links }